% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ej.R
\name{get_ej}
\alias{get_ej}
\title{Get MPCA's Environmental Justice Shapefiles from the MN Geocommons}
\usage{
get_ej(
  path = getwd(),
  folder_name = "mpca_ej_shapes",
  add_date = TRUE,
  use_latlong = FALSE,
  data_only = FALSE
)
}
\arguments{
\item{path}{Directory to save the ZIP file to, if necessary.}

\item{folder_name}{Name of folder to save downloaded ZIP file to, if necessary.}

\item{add_date}{If TRUE, add today's date to the folder name. Helps to prevent overwriting existing folders.}

\item{use_latlong}{Return shapefile with lat/long coordinates (WGS 84). Default is X/Y coordinates in UTM 15N.}

\item{data_only}{Drop the spatial geometry and return a data frame of only the demographic information.}
}
\description{
Download the polygon shapefile of Census Tracts identified by MPCA
as Environmental Justice areas of concern.
}
\details{
Data sources
MN GEOCOMMONS: https://gisdata.mn.gov/
EJ areas: https://gisdata.mn.gov/dataset/env-ej-mpca-census
}
\examples{
library(mpcaej)

\dontrun{
ej <- get_ej()

ej[1, ]

plot(ej[ , 1])
}

}
