% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ej.R
\name{map_ej}
\alias{map_ej}
\title{Map MPCA's Environmental Justice Areas}
\usage{
map_ej()
}
\description{
Map the polygon shapefiles of MN Census Tracts and Tribal Areas identified by MPCA as Environmental Justice areas of concern
}
\details{
Data sources
MN GEOCOMMONS: https://gisdata.mn.gov/
EJ areas: https://gisdata.mn.gov/dataset/env-ej-mpca-census
}
\examples{
library(mpcaej)

\dontrun{

map_ej()

}

}
